Validate poll deletion

+++ src/main/java/cat/xarxarepublicana/hashtagsxrep/application/poll/DeletePollUseCase.java
@@ -1,7 +1,9 @@
 package cat.xarxarepublicana.hashtagsxrep.application.poll;
 
+import cat.xarxarepublicana.hashtagsxrep.domain.error.ValidationException;
 import cat.xarxarepublicana.hashtagsxrep.domain.poll.Poll;
 import cat.xarxarepublicana.hashtagsxrep.domain.poll.PollRepository;
+import org.apache.commons.lang3.StringUtils;
 
 public class DeletePollUseCase {
 
@@ -11,10 +13,14 @@ public class DeletePollUseCase {
         this.pollRepository = pollRepository;
     }
 
-    public void deletePoll(String id) {
-        Poll poll = pollRepository.findById(id);
+    public void deletePoll(String pollId, String descriptionValidator) {
+        Poll poll = pollRepository.findById(pollId);
         if (poll != null) {
-            pollRepository.delete(poll);
+            if (StringUtils.equals(poll.getDescription(), descriptionValidator)) {
+                pollRepository.delete(poll);
+            } else {
+                throw new ValidationException("No coincideixen: [" + poll.getDescription() + "]/[" + descriptionValidator + "]");
+            }
         }
     }
 }
